requirements.txt
fastapi
uvicorn
pydantic
celery
redis
sqlalchemy
psycopg2
python-dotenv
pydantic-settings